requirements.txt
anthropic>=0.93.0
python-dotenv>=1.0.0